warning and dbg hunt

+++ manager/ui/BaseItemUI.h
@@ -58,7 +58,7 @@ public:
 		void paint(juce::Graphics& g)
 		{
 			g.setColour(isMouseOverOrDragging() ? HIGHLIGHT_COLOR : juce::Colours::lightgrey.withAlpha(.3f));
-			for (int i = 0; i < 3; i++) g.drawLine(getWidth() * i / 3, getHeight(), getWidth(), getHeight() * i / 3);
+			for (int i = 0; i < 3; i++) g.drawLine(getWidth() * i / 3.f, (float)getHeight(), (float)getWidth(), getHeight() * i / 3.f);
 		}
 	};
